backend/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-WTF==1.2.*
Flask-Talisman==1.1.*
openai==1.60.*
openpyxl==3.1.*
python-dotenv==1.0.*
